ref: structure of entities files
