--- src/app/common/entities/classes/addedSubjectClass.ts
@@ -1,8 +1,0 @@
-export class AddedSubject {
-  constructor(
-    public subject: string,
-    public teacher: string,
-    public cabiner: number,
-    public desription: string
-  ) {}
-}
