--- src/app/common/entities/classes/index.ts
@@ -1,2 +1,0 @@
-export { AddedSubject } from "./addedSubjectClass";
-export { CreateNewStudent } from "./addedStudentClass";
