+++ src/app/common/entities/index.ts
@@ -1,0 +1,1 @@
+export { CreateNewStudent, IStudent } from "./student";
