--- src/app/common/entities/interfaces/index.ts
@@ -1,3 +1,0 @@
-export { IStudent } from "./studentInterface";
-export { ISubject } from "./subjectInterface";
-export { IRequest } from "./requestInterface";
