--- src/app/common/entities/interfaces/requestInterface.ts
@@ -1,7 +1,0 @@
-export interface IRequest {
-  title: string;
-  firstRow: string;
-  secondRow: string;
-  thirdRow: string;
-  fourthRow: string;
-}
