--- src/app/common/entities/interfaces/studentInterface.ts
@@ -1,8 +1,0 @@
-export interface IStudent {
-  id: number;
-  name: string;
-  lastName: string;
-  address: string;
-  desription: string;
-  subjects: object;
-}
