--- src/app/common/entities/interfaces/subjectInterface.ts
@@ -1,6 +1,0 @@
-export interface ISubject {
-  subject: string;
-  teacher: string;
-  cabiner: number;
-  desription: string;
-}
