+++ tslint.json
@@ -77,7 +77,7 @@
     "no-eval": true,
     "no-inferrable-types": [false],
     "no-internal-module": true,
-    "no-null-keyword": true,
+    "no-null-keyword": false,
     "no-require-imports": true,
     "no-shadowed-variable": true,
     "no-string-literal": true,
